+++ README.md
@@ -42,6 +42,8 @@ Note: Do not modify this section! It is auto-generated by `cobra` using `make ge
 
 - [polycli abi](doc/polycli_abi.md) - Parse an ABI and print the encoded signatures.
 
+- [polycli dbbench](doc/polycli_dbbench.md) - Perform a level/pebble db benchmark
+
 - [polycli dumpblocks](doc/polycli_dumpblocks.md) - Export a range of blocks from a JSON-RPC endpoint.
 
 - [polycli enr](doc/polycli_enr.md) - Convert between ENR and Enode format
@@ -49,8 +51,6 @@ Note: Do not modify this section! It is auto-generated by `cobra` using `make ge
 - [polycli fork](doc/polycli_fork.md) - Take a forked block and walk up the chain to do analysis.
 
 - [polycli hash](doc/polycli_hash.md) - Provide common crypto hashing functions.
-
-- [polycli leveldbbench](doc/polycli_leveldbbench.md) - Perform a level db benchmark
 
 - [polycli loadtest](doc/polycli_loadtest.md) - Run a generic load test against an Eth/EVM style JSON-RPC endpoint.
