+++ cmd/root.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"fmt"
+	"github.com/maticnetwork/polygon-cli/cmd/dbbench"
 	"os"
 
 	"github.com/maticnetwork/polygon-cli/cmd/fork"
@@ -17,7 +18,6 @@ import (
 	"github.com/maticnetwork/polygon-cli/cmd/dumpblocks"
 	"github.com/maticnetwork/polygon-cli/cmd/enr"
 	"github.com/maticnetwork/polygon-cli/cmd/hash"
-	"github.com/maticnetwork/polygon-cli/cmd/leveldbbench"
 	"github.com/maticnetwork/polygon-cli/cmd/loadtest"
 	"github.com/maticnetwork/polygon-cli/cmd/metricsToDash"
 	"github.com/maticnetwork/polygon-cli/cmd/mnemonic"
@@ -104,7 +104,7 @@ func NewPolycliCommand() *cobra.Command {
 		fork.ForkCmd,
 		hash.HashCmd,
 		enr.ENRCmd,
-		leveldbbench.LevelDBBenchCmd,
+		dbbench.DBBenchCmd,
 		loadtest.LoadtestCmd,
 		metricsToDash.MetricsToDashCmd,
 		mnemonic.MnemonicCmd,
