+++ doc/polycli_dbbench.md
@@ -1,3 +1,24 @@
+# `polycli dbbench`
+
+> Auto-generated documentation.
+
+## Table of Contents
+
+- [Description](#description)
+- [Usage](#usage)
+- [Flags](#flags)
+- [See Also](#see-also)
+
+## Description
+
+Perform a level/pebble db benchmark
+
+```bash
+polycli dbbench [flags]
+```
+
+## Usage
+
 This command is meant to give us a sense of the system level
 performance for leveldb:
 
@@ -113,3 +134,47 @@ as well.
 7:58PM DBG recorded result result={"Description":"full scan","EndTime":"2023-07-17T19:58:05.396257711Z","OpCount":9557081144,"OpRate":920614.609547304,"StartTime":"2023-07-17T17:05:04.199777776Z","Stats":{"AliveIterators":0,"AliveSnapshots":0,"BlockCache":{"Buckets":2048,"DelCount":259134854,"GrowCount":9,"HitCount":4,"MissCount":262147633,"Nodes":33294,"SetCount":259168148,"ShrinkCount":2,"Size":268427343},"BlockCacheSize":268427343,"FileCache":{"Buckets":16,"DelCount":536037,"GrowCount":0,"HitCount":2,"MissCount":536537,"Nodes":500,"SetCount":536537,"ShrinkCount":0,"Size":500},"IORead":1092651461848,"IOWrite":13032122717,"Level0Comp":0,"LevelDurations":[0,0,546151937,15675194130,100457643600,40581548153,0],"LevelRead":[0,0,45189458,1233235440,8351239571,3376108236,0],"LevelSizes":[0,103263963,1048356844,10484866671,104856767171,180600915234,797187827055],"LevelTablesCounts":[0,51,665,7066,53522,95777,371946],"LevelWrite":[0,0,45159786,1230799439,8328970986,3371359447,0],"MemComp":0,"NonLevel0Comp":1433,"OpenedTablesCount":500,"SeekComp":0,"WriteDelayCount":0,"WriteDelayDuration":0,"WritePaused":false},"TestDuration":10381196479925,"ValueDist":null}
 
 ```
+
+## Flags
+
+```bash
+      --cache-size int                the number of megabytes to use as our internal cache size (default 512)
+      --db-mode string                The mode to use: leveldb or pebbledb (default "leveldb")
+      --db-path string                the path of the database that we'll use for testing (default "_benchmark_db")
+      --degree-of-parallelism uint8   The number of concurrent goroutines we'll use (default 2)
+      --dont-fill-read-cache          if false, then random reads will be cached
+      --full-scan-mode                if true, the application will scan the full database as fast as possible and print a summary
+      --handles int                   defines the capacity of the open files caching. Use -1 for zero, this has same effect as specifying NoCacher to OpenFilesCacher. (default 500)
+  -h, --help                          help for dbbench
+      --key-size uint                 The byte length of the keys that we'll use (default 32)
+      --nil-read-opts                 if true we'll use nil read opt (this is what geth/bor does)
+      --no-merge-write                allows disabling write merge
+      --overwrite-count uint          the number of times to overwrite the data (default 5)
+      --read-limit uint               the number of reads will attempt to complete in a given test (default 10000000)
+      --read-only                     if true, we'll skip all the write operations and open the DB in read only mode
+      --read-strict                   if true the rand reads will be made in strict mode
+      --sequential-reads              if true we'll perform reads sequentially
+      --sequential-writes             if true we'll perform writes in somewhat sequential manner
+      --size-distribution string      the size distribution to use while testing (default "0-1:2347864,2-3:804394856,4-7:541267689,8-15:738828593,16-31:261122372,32-63:1063470933,64-127:3584745195,128-255:1605760137,256-511:316074206,512-1023:312887514,1024-2047:328894149,2048-4095:141180,4096-8191:92789,8192-16383:256060,16384-32767:261806,32768-65535:191032,65536-131071:99715,131072-262143:73782,262144-524287:17552,524288-1048575:717,1048576-2097151:995,2097152-4194303:1,8388608-16777215:1")
+      --sync-writes                   sync each write
+      --write-limit uint              The number of entries to write in the db (default 1000000)
+      --write-zero                    if true, we'll write 0s rather than random data
+```
+
+The command also inherits flags from parent commands.
+
+```bash
+      --config string   config file (default is $HOME/.polygon-cli.yaml)
+      --pretty-logs     Should logs be in pretty format or JSON (default true)
+  -v, --verbosity int   0 - Silent
+                        100 Fatal
+                        200 Error
+                        300 Warning
+                        400 Info
+                        500 Debug
+                        600 Trace (default 400)
+```
+
+## See also
+
+- [polycli](polycli.md) - A Swiss Army knife of blockchain tools.
